+++ 06-Code/MilitaryDiningHall2/MilitaryDiningHall/src/ec/edu/espe/militarydininghall/model/Commensal.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package ec.edu.espe.militarydininghall.model;
 
 import com.google.gson.Gson;
@@ -16,7 +12,7 @@ import utils.Validations;
  */
 public class Commensal {
 
-    private int id;
+    private long id;
     private String name;
     private String email;
     private String password;
@@ -24,7 +20,7 @@ public class Commensal {
     private String type;
     private float balance;
 
-    public Commensal(int id, String name, String email, String password, String grade, String type, float balance) {
+    public Commensal(long id, String name, String email, String password, String grade, String type, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -34,10 +30,12 @@ public class Commensal {
         this.balance = balance;
     }
 
+
+
     /**
      * @return the id
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
